rename of constants namespace in Repository
