+++ source/Energinet.DataHub.PostOffice.Infrastructure/Repositories/DataAvailableNotificationCleanUpRepository.cs
@@ -17,8 +17,8 @@ using System.Threading.Tasks;
 using Energinet.DataHub.PostOffice.Domain.Repositories;
 using Energinet.DataHub.PostOffice.Infrastructure.Common;
 using Energinet.DataHub.PostOffice.Infrastructure.Documents;
-using Energinet.DataHub.PostOffice.Infrastructure.Repositories.Constants;
 using Energinet.DataHub.PostOffice.Infrastructure.Repositories.Containers;
+using Energinet.DataHub.PostOffice.Infrastructure.Repositories.Helpers;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Extensions.Logging;
 using NodaTime;
