+++ source/Energinet.DataHub.PostOffice.IntegrationTests/Repositories/DataAvailableNotificationCleanUpRepositoryTests.cs
@@ -19,8 +19,8 @@ using System.Threading.Tasks;
 using Energinet.DataHub.PostOffice.Domain.Model;
 using Energinet.DataHub.PostOffice.Domain.Repositories;
 using Energinet.DataHub.PostOffice.Infrastructure.Documents;
-using Energinet.DataHub.PostOffice.Infrastructure.Repositories.Constants;
 using Energinet.DataHub.PostOffice.Infrastructure.Repositories.Containers;
+using Energinet.DataHub.PostOffice.Infrastructure.Repositories.Helpers;
 using Energinet.DataHub.PostOffice.IntegrationTests.Common;
 using Microsoft.Azure.Cosmos;
 using Moq;
